simpler nav, /logout route for logging out
