+++ libs/frontend/src/lib/config/general.ts
@@ -1,2 +1,2 @@
 export const githubRepo = "https://github.com/juicemitapfelndrin/codincod";
-export const websiteName = "Codin Cod";
+export const websiteName = "CodinCod";
