+++ libs/frontend/src/routes/logout/+page.server.ts
@@ -1,0 +1,6 @@
+import { cookieKeys, frontendUrls } from "types";
+
+export const load = async ({ cookies }) => {
+	cookies.delete(cookieKeys.TOKEN, { path: frontendUrls.ROOT });
+	return { loggedOut: true };
+};
